+++ mvc/src/main/java/web/org/springframework/web/support/web/filter/ResourceFilter.java
@@ -1,4 +1,4 @@
-package com.techcourse.support.web.filter;
+package web.org.springframework.web.support.web.filter;
 
 import jakarta.servlet.*;
 import jakarta.servlet.annotation.WebFilter;
